AP_HAL_SITL: uart fix implicit cast and pb of different type

+++ libraries/AP_HAL_SITL/UARTDriver.cpp
@@ -58,17 +58,17 @@ void UARTDriver::begin(uint32_t baud, uint16_t rxSpace, uint16_t txSpace)
         char *args1 = strtok_r(nullptr, ":", &saveptr);
         char *args2 = strtok_r(nullptr, ":", &saveptr);
         if (strcmp(devtype, "tcp") == 0) {
-            uint16_t port = atoi(args1);
+            uint16_t port = static_cast<uint16_t>(atoi(args1));
             const bool wait = (args2 && strcmp(args2, "wait") == 0);
             _tcp_start_connection(port, wait);
         } else if (strcmp(devtype, "tcpclient") == 0) {
             if (args2 == nullptr) {
                 AP_HAL::panic("Invalid tcp client path: %s", path);
             }
-            const uint16_t port = atoi(args2);
+            const uint16_t port = static_cast<uint16_t>(atoi(args2));
             _tcp_start_client(args1, port);
         } else if (strcmp(devtype, "uart") == 0) {
-            const uint32_t baudrate = args2 ? atoi(args2) : baud;
+            const uint32_t baudrate = args2 ? static_cast<uint32_t>(atoi(args2)) : baud;
             ::printf("UART connection %s:%u\n", args1, baudrate);
             _uart_path = strdup(args1);
             _uart_baudrate = baudrate;
@@ -373,7 +373,7 @@ bool UARTDriver::_select_check(int fd)
 
 void UARTDriver::_set_nonblocking(int fd)
 {
-    unsigned v = fcntl(fd, F_GETFL, 0);
+    int32_t v = fcntl(fd, F_GETFL, 0);
     fcntl(fd, F_SETFL, v | O_NONBLOCK);
 }
